requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
numpy==1.19.4
pkg-resources==0.0.0
pylint==2.6.0
python-dotenv==0.15.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.23
style==1.1.0
toml==0.10.2
update==0.0.1
wrapt==1.12.1
